+++ db/diff.sql
@@ -1,2 +1,7 @@
--- This is an empty migration.
+-- AlterTable
+ALTER TABLE "MarkdownPage" ALTER COLUMN "githubSha" DROP NOT NULL;
+
+-- AlterTable
+ALTER TABLE "SiteBranch" DROP COLUMN "trieveDatasetId",
+DROP COLUMN "trieveReadApiKey";